Duc v2

+++ frontend/src/containers/Cart/index.tsx
@@ -7,23 +7,37 @@ import { cartListAction } from '../../redux/cart/cart';
 import { Link } from 'react-router-dom';
 import { UIState } from '../../common/interfaces/UI';
 import { UIListAction } from '../../redux/UI/UI';
-import { OrderItemToSend, OrderListToSend } from '../../common/interfaces/orderList';
+import { OrderItemToSend, OrderListToSend, ShipmentDetailDTO } from '../../common/interfaces/orderList';
 import { orderListThunk } from '../../redux/orderList/orderListThunk';
 import MinusIcon from '../../components/common/icon/Minus';
 import AddIcon from '../../components/common/icon/Add';
+import InputField from '../../components/common/InputField';
+import { useForm } from 'react-hook-form';
+import { UserState } from '../../common/interfaces/user';
 
 interface CartProps {}
 
 const Cart: React.FunctionComponent<CartProps> = () => {
+    const { handleSubmit, register, setValue } = useForm<ShipmentDetailDTO>();
     const cartState = useSelector<RootState, CartState>((state) => state.cart);
     const UIState = useSelector<RootState, UIState>((state) => state.UI);
-    const [totalPrice, setTotalPrice] = React.useState(0);
-    //
+    const userState = useSelector<RootState, UserState>((state) => state.user);
+    const [totalPrice, setTotalPrice] = React.useState<number>(0);
+    const [isConfirm, setIsConfirm] = React.useState<boolean>(false);
+    //set default value on first rendered
     React.useEffect(() => {
-        if (UIState.confirmPopUp.isConfirm) {
-            store.dispatch(cartListAction.deleteProduct(UIState.confirmPopUp.productToDelete));
+        setValue('address', userState.user.address);
+        setValue('phoneNumber', userState.user.phoneNumber);
+    }, [userState.user.address, userState.user.phoneNumber, setValue]);
+
+    //action when delete confirm popup appear
+    React.useEffect(() => {
+        if (UIState.confirmPopUp.isConfirm != null) {
+            if (UIState.confirmPopUp.isConfirm) {
+                store.dispatch(cartListAction.deleteProduct(UIState.confirmPopUp.productToDelete));
+            }
+            store.dispatch(UIListAction.resetConfirmPopup());
         }
-        store.dispatch(UIListAction.resetConfirmPopup());
     }, [UIState.confirmPopUp.isConfirm, UIState.confirmPopUp.productToDelete]);
     //recalculate the total price when a product in cart was changed
     React.useEffect(() => {
@@ -34,8 +48,8 @@ const Cart: React.FunctionComponent<CartProps> = () => {
         );
     }, [cartState.productList]);
 
-    //
-    const onHandleOrderClick = async () => {
+    //on submit form
+    const onSubmit = async (data: ShipmentDetailDTO) => {
         const orderItemList = cartState.productList.map<OrderItemToSend>((item) => {
             return {
                 amount: item.quantity,
@@ -47,6 +61,8 @@ const Cart: React.FunctionComponent<CartProps> = () => {
         const orderListToSend: OrderListToSend = {
             status: 'WAITING',
             orderItem: orderItemList,
+            address: data.address,
+            phoneNumber: data.phoneNumber,
         };
         const result = await store.dispatch(orderListThunk.addOrderList(orderListToSend));
         if (result.meta.requestStatus === 'fulfilled') {
@@ -80,7 +96,10 @@ const Cart: React.FunctionComponent<CartProps> = () => {
                         <h1 className="text-3xl font-extrabold tracking-tight text-gray-900 sm:text-4xl">
                             Shopping Cart
                         </h1>
-                        <form className="mt-12 lg:grid lg:grid-cols-12 lg:gap-x-12 lg:items-start xl:gap-x-16">
+                        <form
+                            onSubmit={handleSubmit(onSubmit)}
+                            className="mt-12 lg:grid lg:grid-cols-12 lg:gap-x-12 lg:items-start xl:gap-x-16"
+                        >
                             <section aria-labelledby="cart-heading" className="lg:col-span-7">
                                 <h2 id="cart-heading" className="sr-only">
                                     Items in your shopping cart
@@ -205,38 +224,85 @@ const Cart: React.FunctionComponent<CartProps> = () => {
                                 aria-labelledby="summary-heading"
                                 className="px-4 py-6 mt-16 rounded-lg bg-gray-50 sm:p-6 lg:p-8 lg:mt-0 lg:col-span-5 intro-y"
                             >
-                                <h2 id="summary-heading" className="text-lg font-medium text-gray-900">
-                                    Order summary
-                                </h2>
+                                <div className="">
+                                    <h2 id="summary-heading" className="text-lg font-bold text-gray-900">
+                                        Order summary
+                                    </h2>
 
-                                <dl className="mt-6 space-y-4">
-                                    <div className="flex items-center justify-between">
-                                        <dt className="text-sm text-gray-600">Subtotal</dt>
-                                        <dd className="text-sm font-medium text-gray-900">{totalPrice}$</dd>
+                                    <dl className="mt-6 space-y-4">
+                                        <div className="flex items-center justify-between">
+                                            <dt className="text-sm text-gray-600">Subtotal</dt>
+                                            <dd className="text-sm font-medium text-gray-900">{totalPrice}$</dd>
+                                        </div>
+                                        <div className="flex items-center justify-between pt-4 border-t border-gray-200">
+                                            <dt className="flex items-center text-sm text-gray-600">
+                                                <span>Shipping estimate</span>
+                                            </dt>
+                                            <dd className="text-sm font-medium text-gray-900">-</dd>
+                                        </div>
+                                        <div className="flex items-center justify-between pt-4 border-t border-gray-200">
+                                            <dt className="flex text-sm text-gray-600">
+                                                <span>Tax estimate</span>
+                                            </dt>
+                                            <dd className="text-sm font-medium text-gray-900">-</dd>
+                                        </div>
+                                        <div className="flex items-center justify-between pt-4 border-t border-gray-200">
+                                            <dt className="text-base font-medium text-gray-900">Order total</dt>
+                                            <dd className="text-base font-medium text-gray-900">{totalPrice}$</dd>
+                                        </div>
+                                    </dl>
+                                </div>
+
+                                <div className="mt-5 ">
+                                    <h2 id="summary-heading" className="text-lg font-bold text-gray-900">
+                                        Shipment detail
+                                    </h2>
+                                    <div className="mt-5">
+                                        <InputField
+                                            type="text"
+                                            label="Address"
+                                            field="address"
+                                            defaultValue={userState.user.address}
+                                            message={''}
+                                            register={register}
+                                        />
                                     </div>
-                                    <div className="flex items-center justify-between pt-4 border-t border-gray-200">
-                                        <dt className="flex items-center text-sm text-gray-600">
-                                            <span>Shipping estimate</span>
-                                        </dt>
-                                        <dd className="text-sm font-medium text-gray-900">-</dd>
+                                    <div className="mt-5">
+                                        <InputField
+                                            type="text"
+                                            label="Phone number"
+                                            field="phoneNumber"
+                                            defaultValue={userState.user.phoneNumber}
+                                            message={''}
+                                            register={register}
+                                        />
                                     </div>
-                                    <div className="flex items-center justify-between pt-4 border-t border-gray-200">
-                                        <dt className="flex text-sm text-gray-600">
-                                            <span>Tax estimate</span>
-                                        </dt>
-                                        <dd className="text-sm font-medium text-gray-900">-</dd>
-                                    </div>
-                                    <div className="flex items-center justify-between pt-4 border-t border-gray-200">
-                                        <dt className="text-base font-medium text-gray-900">Order total</dt>
-                                        <dd className="text-base font-medium text-gray-900">{totalPrice}$</dd>
-                                    </div>
-                                </dl>
+                                </div>
 
                                 <div className="mt-6">
+                                    <div className="flex items-center mb-3">
+                                        <input
+                                            id="confirm-order"
+                                            name="confirm-order"
+                                            type="checkbox"
+                                            checked={isConfirm}
+                                            onChange={(e) => {
+                                                setIsConfirm(e.target.checked);
+                                            }}
+                                            className="w-4 h-4 text-indigo-600 border-gray-300 rounded focus:ring-indigo-500"
+                                        />
+                                        <label htmlFor="confirm-order" className="block ml-2 text-sm text-gray-900">
+                                            I accept to order these product with this shipment detail
+                                        </label>
+                                    </div>
                                     <button
-                                        type="button"
-                                        className="w-full px-4 py-3 text-base font-medium text-white bg-indigo-600 border border-transparent rounded-md shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-50 focus:ring-indigo-500"
-                                        onClick={onHandleOrderClick}
+                                        type="submit"
+                                        className={
+                                            isConfirm
+                                                ? 'w-full px-4 py-3 text-base font-medium text-white bg-indigo-600 border border-transparent rounded-md shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-50 focus:ring-indigo-500'
+                                                : 'w-full px-4 py-3 text-base font-medium text-white bg-gray-600 border border-transparent rounded-md shadow-sm cursor-not-allowed'
+                                        }
+                                        disabled={!isConfirm}
                                     >
                                         Order now
                                     </button>
